fix: web share api util로 변경
